Format MACDCross hyperopt
